+++ app/src/main/java/com/itssuryansh/taaveez/domain/use_cases/CreateNoteUseCase.kt
@@ -1,0 +1,11 @@
+package com.itssuryansh.taaveez.domain.use_cases
+
+import com.itssuryansh.taaveez.domain.model.NotesEntity
+import com.itssuryansh.taaveez.domain.repository.NotesRepository
+import javax.inject.Inject
+
+class CreateNoteUseCase @Inject constructor (private val notesRepository: NotesRepository){
+    suspend operator fun invoke(notesEntity: NotesEntity){
+        notesRepository.insert(notesEntity)
+    }
+}
