+++ app/src/main/java/com/itssuryansh/taaveez/utilities/Constants.kt
@@ -1,4 +1,4 @@
-package com.itssuryansh.taaveez
+package com.itssuryansh.taaveez.utilities
 
 object Constants {
