+++ app/src/main/res/layout/activity_add_new_content.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/bg"
-    tools:context="com.itssuryansh.taaveez.Add_New_Content">
+    tools:context="com.itssuryansh.taaveez.presentation.ui.Add_New_Content">
 
 
     <FrameLayout
