+++ app/src/main/res/layout/activity_edit_content.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/bg"
-    tools:context=".Edit_Content">
+    tools:context=".presentation.ui.Edit_Content">
 
 
     <FrameLayout
