+++ build.gradle
@@ -13,6 +13,8 @@ plugins {
     id 'com.android.application' version '7.4.2' apply false
     id 'com.android.library' version '7.4.2' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.20' apply false
+    id("com.google.dagger.hilt.android") version "2.44" apply false
+
 
 }
